+++ NEWS.md
@@ -1,7 +1,10 @@
 # timemoir 0.3.0.9000
 
-* allow benchmarking of multiple functions using ellipsis
-* the max_mem returned is now max_mem - min_mem
+* allow benchmarking of multiple functions using ellipsis.
+
+* the max_mem returned is now max_mem - min_mem.
+
+* add a verbose parameter.
 
 # timemoir 0.2.0.9000